move folder into workspace
