+++ src/packages/user/user-group/manifests.ts
@@ -4,7 +4,7 @@ import { manifests as entityBulkActionManifests } from './entity-bulk-actions/ma
 import { manifests as menuItemManifests } from './menu-item/manifests.js';
 import { manifests as modalManifests } from './modals/manifests.js';
 import { manifests as repositoryManifests } from './repository/manifests.js';
-import { manifests as sectionViewManifests } from './section-view/manifests.js';
+import { manifests as sectionViewManifests } from './workspace/user-group-root/manifests.js';
 import { manifests as workspaceManifests } from './workspace/manifests.js';
 
 import type { ManifestTypes, UmbBackofficeManifestKind } from '@umbraco-cms/backoffice/extension-registry';
