+++ src/packages/user/user-group/workspace/user-group-root/manifests.ts
@@ -1,4 +1,4 @@
-import { UMB_USER_GROUP_ROOT_ENTITY_TYPE } from '../entity.js';
+import { UMB_USER_GROUP_ROOT_ENTITY_TYPE } from '../../entity.js';
 import type { ManifestTypes, UmbBackofficeManifestKind } from '@umbraco-cms/backoffice/extension-registry';
 
 export const manifests: Array<ManifestTypes | UmbBackofficeManifestKind> = [
