+++ src/packages/user/user-group/workspace/user-group-root/user-group-root-workspace.element.ts
@@ -1,4 +1,4 @@
-import { UMB_USER_GROUP_COLLECTION_ALIAS } from '../collection/index.js';
+import { UMB_USER_GROUP_COLLECTION_ALIAS } from '../../collection/index.js';
 import { html, customElement } from '@umbraco-cms/backoffice/external/lit';
 import { UmbLitElement } from '@umbraco-cms/backoffice/lit-element';
